requirements.txt
PyQt6>=6.10.0
numpy>=2.3.4
Pillow>=12.0.0
